+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # v1.0.6 - TBD
 
+## Added
+
+- [#235](https://github.com/hyperf-cloud/hyperf/pull/235) Added default exception handler and optimized code.
 
 # v1.0.5 - 2019-07-07